LPS-86168 Moves code to a new class
